requirements.txt
fastapi
uvicorn
torch
torchvision
pandas
scikit-learn
pillow
